src/main/java/com/switchfully/euder/service/ItemService.java
package com.switchfully.euder.service;

import com.switchfully.euder.domain.entities.item.Item;
import com.switchfully.euder.domain.repositories.ItemRepository;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

@Component
public class ItemService {
    private static final Logger LOGGER = LoggerFactory.getLogger(ItemService.class);
    private final ItemRepository itemRepository;

    @Autowired
    public ItemService(ItemRepository itemRepository) {
        this.itemRepository = itemRepository;
    }

    public Item createItem(Item item) {
        LOGGER.info("Inserting one Item");
        return itemRepository.createItem(item);
    }
}
